feat: rename to brush-circular
