+++ src/tracks/brush-circular/index.ts
@@ -1,0 +1,2 @@
+export { CircularBrushTrack } from './brush-circular-plot';
+export type { CircularBrushTrackOptions, CircularBrushTrackContext } from './brush-circular';
